Renamed loading-state.ts file.
