+++ src/openssf-dashboard/shared/loading-state.ts
